video/hd61602.cpp: Moving HD61602 to its own file.
